refactor(useSelected): useSelected 분리
